requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
click==8.1.8
eval-type-backport==0.2.2
exceptiongroup==1.3.0
faiss-cpu==1.7.4
flask==3.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
invoke==2.2.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
mistralai==1.9.10
numpy==1.26.0
orjson==3.11.3
pydantic==2.11.7
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
typing-extensions==4.15.0
typing-inspection==0.4.1
werkzeug==3.0.3
zipp==3.23.0
